Remove Pipe (Processor) logic from ESQL-core
