+++ x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/expression/function/scalar/BaseSurrogateFunction.java
@@ -8,7 +8,6 @@
 package org.elasticsearch.xpack.esql.core.expression.function.scalar;
 
 import org.elasticsearch.xpack.esql.core.expression.Expression;
-import org.elasticsearch.xpack.esql.core.expression.gen.pipeline.Pipe;
 import org.elasticsearch.xpack.esql.core.tree.Source;
 
 import java.util.List;
@@ -45,8 +44,4 @@ public abstract class BaseSurrogateFunction extends ScalarFunction implements Su
         return substitute().fold();
     }
 
-    @Override
-    protected Pipe makePipe() {
-        return substitute().asPipe();
-    }
 }
