+++ x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/expression/predicate/Range.java
@@ -7,14 +7,8 @@
 package org.elasticsearch.xpack.esql.core.expression.predicate;
 
 import org.elasticsearch.xpack.esql.core.expression.Expression;
-import org.elasticsearch.xpack.esql.core.expression.Expressions;
 import org.elasticsearch.xpack.esql.core.expression.function.scalar.ScalarFunction;
-import org.elasticsearch.xpack.esql.core.expression.gen.pipeline.Pipe;
-import org.elasticsearch.xpack.esql.core.expression.predicate.logical.BinaryLogicPipe;
-import org.elasticsearch.xpack.esql.core.expression.predicate.logical.BinaryLogicProcessor.BinaryLogicOperation;
 import org.elasticsearch.xpack.esql.core.expression.predicate.operator.comparison.BinaryComparison;
-import org.elasticsearch.xpack.esql.core.expression.predicate.operator.comparison.BinaryComparisonPipe;
-import org.elasticsearch.xpack.esql.core.expression.predicate.operator.comparison.BinaryComparisonProcessor.BinaryComparisonOperation;
 import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
 import org.elasticsearch.xpack.esql.core.tree.Source;
 import org.elasticsearch.xpack.esql.core.type.DataType;
@@ -134,26 +128,6 @@ public class Range extends ScalarFunction {
     @Override
     public DataType dataType() {
         return DataTypes.BOOLEAN;
-    }
-
-    @Override
-    protected Pipe makePipe() {
-        BinaryComparisonPipe lowerPipe = new BinaryComparisonPipe(
-            source(),
-            this,
-            Expressions.pipe(value()),
-            Expressions.pipe(lower()),
-            includeLower() ? BinaryComparisonOperation.GTE : BinaryComparisonOperation.GT
-        );
-        BinaryComparisonPipe upperPipe = new BinaryComparisonPipe(
-            source(),
-            this,
-            Expressions.pipe(value()),
-            Expressions.pipe(upper()),
-            includeUpper() ? BinaryComparisonOperation.LTE : BinaryComparisonOperation.LT
-        );
-        BinaryLogicPipe and = new BinaryLogicPipe(source(), this, lowerPipe, upperPipe, BinaryLogicOperation.AND);
-        return and;
     }
 
     @Override
