+++ x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/expression/predicate/operator/comparison/BinaryComparison.java
@@ -7,10 +7,8 @@
 package org.elasticsearch.xpack.esql.core.expression.predicate.operator.comparison;
 
 import org.elasticsearch.xpack.esql.core.expression.Expression;
-import org.elasticsearch.xpack.esql.core.expression.Expressions;
 import org.elasticsearch.xpack.esql.core.expression.TypeResolutions;
 import org.elasticsearch.xpack.esql.core.expression.TypeResolutions.ParamOrdinal;
-import org.elasticsearch.xpack.esql.core.expression.gen.pipeline.Pipe;
 import org.elasticsearch.xpack.esql.core.expression.predicate.BinaryOperator;
 import org.elasticsearch.xpack.esql.core.expression.predicate.operator.comparison.BinaryComparisonProcessor.BinaryComparisonOperation;
 import org.elasticsearch.xpack.esql.core.tree.Source;
@@ -41,11 +39,6 @@ public abstract class BinaryComparison extends BinaryOperator<Object, Object, Bo
     @Override
     public DataType dataType() {
         return DataTypes.BOOLEAN;
-    }
-
-    @Override
-    protected Pipe makePipe() {
-        return new BinaryComparisonPipe(source(), this, Expressions.pipe(left()), Expressions.pipe(right()), function());
     }
 
     public static Integer compare(Object left, Object right) {
