+++ x-pack/plugin/esql-core/src/test/java/org/elasticsearch/xpack/esql/core/optimizer/OptimizerRulesTests.java
@@ -14,7 +14,6 @@ import org.elasticsearch.xpack.esql.core.expression.Expressions;
 import org.elasticsearch.xpack.esql.core.expression.FieldAttribute;
 import org.elasticsearch.xpack.esql.core.expression.Literal;
 import org.elasticsearch.xpack.esql.core.expression.Nullability;
-import org.elasticsearch.xpack.esql.core.expression.function.scalar.string.StartsWith;
 import org.elasticsearch.xpack.esql.core.expression.predicate.BinaryOperator;
 import org.elasticsearch.xpack.esql.core.expression.predicate.Predicates;
 import org.elasticsearch.xpack.esql.core.expression.predicate.Range;
@@ -1745,50 +1744,6 @@ public class OptimizerRulesTests extends ESTestCase {
         Filter expected = new Filter(EMPTY, relation, new And(EMPTY, new And(EMPTY, isNotNull(fieldA), isNotNull(fieldB)), inn));
 
         assertEquals(expected, new OptimizerRules.InferIsNotNull().apply(f));
-    }
-
-    public void testIsNotNullOnFunctionWithOneField() {
-        EsRelation relation = relation();
-        var fieldA = TestUtils.getFieldAttribute("a");
-        var pattern = L("abc");
-        Expression inn = isNotNull(
-            new And(EMPTY, new TestStartsWith(EMPTY, fieldA, pattern, false), greaterThanOf(new Add(EMPTY, ONE, TWO), THREE))
-        );
-
-        Filter f = new Filter(EMPTY, relation, inn);
-        Filter expected = new Filter(EMPTY, relation, new And(EMPTY, isNotNull(fieldA), inn));
-
-        assertEquals(expected, new OptimizerRules.InferIsNotNull().apply(f));
-    }
-
-    public void testIsNotNullOnFunctionWithTwoFields() {
-        EsRelation relation = relation();
-        var fieldA = TestUtils.getFieldAttribute("a");
-        var fieldB = TestUtils.getFieldAttribute("b");
-        var pattern = L("abc");
-        Expression inn = isNotNull(new TestStartsWith(EMPTY, fieldA, fieldB, false));
-
-        Filter f = new Filter(EMPTY, relation, inn);
-        Filter expected = new Filter(EMPTY, relation, new And(EMPTY, new And(EMPTY, isNotNull(fieldA), isNotNull(fieldB)), inn));
-
-        assertEquals(expected, new OptimizerRules.InferIsNotNull().apply(f));
-    }
-
-    public static class TestStartsWith extends StartsWith {
-
-        public TestStartsWith(Source source, Expression input, Expression pattern, boolean caseInsensitive) {
-            super(source, input, pattern, caseInsensitive);
-        }
-
-        @Override
-        public Expression replaceChildren(List<Expression> newChildren) {
-            return new TestStartsWith(source(), newChildren.get(0), newChildren.get(1), isCaseInsensitive());
-        }
-
-        @Override
-        protected NodeInfo<TestStartsWith> info() {
-            return NodeInfo.create(this, TestStartsWith::new, input(), pattern(), isCaseInsensitive());
-        }
     }
 
     public void testIsNotNullOnFunctionWithTwoField() {}
